Struct simple_vector2d::Vector2

pub struct Vector2<T>(pub T, pub T);

Representation of a mathematical vector e.g. a position or velocity

Methods

impl<T: Float> Vector2<T>

fn unit_vector(direction: T) -> Self

Creates a new unit vector in a specific direction

fn normalise(self) -> Self

Normalises the vector

fn length(self) -> T

Returns the magnitude/length of the vector

fn length_squared(self) -> T

Returns the magnitude/length of the vector squared

fn direction(self) -> T

Returns direction the vector is pointing

fn direction_to(self, other: Self) -> T

Returns direction towards another vector

fn distance_to(self, other: Self) -> T

Returns the distance betweens two vectors

fn distance_to_squared(self, other: Self) -> T

Returns the distance betweens two vectors

fn is_any_nan(&self) -> bool

Returns true if either component is NaN.

fn is_any_infinite(&self) -> bool

Returns true if either component is positive or negative infinity.

fn is_all_finite(&self) -> bool

Returns true if both components are neither infinite nor NaN.

fn is_all_normal(&self) -> bool

Returns true if both components are neither zero, infinite, subnormal nor NaN.

impl<T> Vector2<T>

fn dot(self, other: Self) -> T::Output::Output where T: Mul, T::Output: Add

Returns the dot product of two vectors